temp

+++ src/main.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 vector<tech> tech_stack;
-int die_area;
+long long int die_area;
 int top_die_max_util, bottom_die_max_util;
 int top_repeat_count, bottom_repeat_count;
